requirements.txt
psutil==5.8.0
apiclient==1.0.4
beautifulsoup4==4.10.0
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
httplib2==0.20.2
oauth2client==4.1.3
python-dotenv==0.19.2
google-auth-oauthlib==0.4.6
selenium==3.141.0
pytest==6.2.5
pytest-xdist
